重大重构 1. 由于ASE-Ramake临近发布，存在重大不兼容更改，弃用ASE Dev Tool插件，转为使用ASER Studio。新的ASER Studio将提供完整文档。 2. 重命名VISPM为VIPLM以访问插件管理器实例 3. 添加VIPM和VIPlaceholder宏用于访问占位符功能 4. 补全JsonConfig遗漏的const语义 5. 全面弃用PathMacro 6. 初步实现DragWidget 7. 变更YSS::MainWin生命周期策略，解决一些资源释放和使用不正确的问题 8. 暂时移除YSS::MainWin的样式表，等待ASERStudio语法完成后重新开发样式 9. 为YSSCore::FileServer增加特别关注功能

+++ src/Visindigo/General/private/LoggerMsgHandler_impl.hpp
@@ -35,7 +35,7 @@ namespace Visindigo::General {
 		QString temp = "QList[";
 		for (int i = 0; i < list.size(); i++) {
 			void* ptr = &list[i];
-			temp += " [Object at " % QString::number(ptr, 16) % "],";
+			temp += " [Object at " % QString::number((quint64)ptr, 16) % "],";
 		}
 		temp.removeLast();
 		temp += "}";
